split up into header and implementation file, added documentation, a doxyfile and a makefile
